Create extension for AVCaptureVideoOriendation / currentOrientation prop
